FIX Removing the individual task for this package and instead extending the main task in silverstripe-maintenance
