+++ src/Extensions/ComposerLoaderExtension.php
@@ -5,6 +5,10 @@ namespace BringYourOwnIdeas\UpdateChecker\Extensions;
 use Composer\Composer;
 use Composer\Factory;
 use Composer\IO\NullIO;
+use Composer\Package\Link;
+use Composer\Repository\ArrayRepository;
+use Composer\Repository\BaseRepository;
+use Composer\Repository\CompositeRepository;
 use Extension;
 
 class ComposerLoaderExtension extends Extension
@@ -30,6 +34,64 @@ class ComposerLoaderExtension extends Extension
     public function getComposer()
     {
         return $this->composer;
+    }
+
+    /**
+     * Retrieve an array of primary composer dependencies from composer.json.
+     *
+     * Packages are filtered by allowed type.
+     *
+     * @param array|null $allowedTypes An array of "allowed" package types. Dependencies in composer.json that do not
+     *                                 match any of the given types are not returned.
+     * @return array[]
+     */
+    public function getPackages(array $allowedTypes = null)
+    {
+        /** @var Composer $composer */
+        $composer = $this->getComposer();
+
+        /** @var BaseRepository $repository */
+        $repository = new CompositeRepository([
+            new ArrayRepository([$composer->getPackage()]),
+            $composer->getRepositoryManager()->getLocalRepository(),
+        ]);
+
+        $packages = [];
+        foreach ($repository->getPackages() as $package) {
+            // Filter out packages that are not "allowed types"
+            if (is_array($allowedTypes) && !in_array($package->getType(), $allowedTypes)) {
+                continue;
+            }
+
+            // Find the constraint used for installation
+            $constraint = $this->getInstalledConstraint($repository, $package->getName());
+            $packages[$package->getName()] = [
+                'constraint' => $constraint,
+                'package' => $package,
+            ];
+        }
+        return $packages;
+    }
+
+    /**
+     * Find all dependency constraints for the given package in the current repository and return the strictest one
+     *
+     * @param BaseRepository $repository
+     * @param string $packageName
+     * @return string
+     */
+    protected function getInstalledConstraint(BaseRepository $repository, $packageName)
+    {
+        $constraints = [];
+        foreach ($repository->getDependents($packageName) as $dependent) {
+            /** @var Link $link */
+            list (, $link) = $dependent;
+            $constraints[] = $link->getPrettyConstraint();
+        }
+
+        usort($constraints, 'version_compare');
+
+        return array_pop($constraints);
     }
 
     /**
